Refactor the tests of StocksPortfolio/index.test.js file and create an testing-utils.js for it
